move create_test_extraction_data to test_utils.py
