+++ ark/utils/segmentation_utils_test.py
@@ -6,33 +6,6 @@ from skimage.measure import regionprops
 from ark.utils import segmentation_utils
 
 
-def _create_test_extraction_data():
-    # first create segmentation masks
-    cell_mask = np.zeros((40, 40), dtype='int16')
-    cell_mask[4:10, 4:10] = 1
-    cell_mask[15:25, 20:30] = 2
-    cell_mask[27:32, 3:28] = 3
-    cell_mask[35:40, 15:22] = 5
-
-    # then create channels data
-    channel_data = np.zeros((40, 40, 5), dtype="int16")
-    channel_data[:, :, 0] = 1
-    channel_data[:, :, 1] = 5
-    channel_data[:, :, 2] = 5
-    channel_data[:, :, 3] = 10
-    channel_data[:, :, 4] = 0
-
-    # cell1 is the only cell negative for channel 3
-    cell1 = cell_mask == 1
-    channel_data[cell1, 3] = 0
-
-    # cell2 is the only cell positive for channel 4
-    cell2 = cell_mask == 2
-    channel_data[cell2, 4] = 10
-
-    return cell_mask, channel_data
-
-
 def test_find_nuclear_mask_id():
     # create cell labels with 5 distinct cells
     cell_labels = np.zeros((60, 10), dtype='int')
